# Conflicts:
#	data/src/main/java/com/cheesecake/data/repository/IFootballRepositoryImpl.kt
#	domain/src/main/java/com/cheesecake/domain/repository/IFootballRepository.kt
#	presentation/src/main/res/navigation/nav_graph.xml
#	presentation/src/main/res/values/strings.xml

+++ data/src/main/java/com/cheesecake/data/remote/models/MatchesTotalDTO.kt
@@ -3,7 +3,7 @@ package com.cheesecake.data.remote.models
 import com.google.gson.annotations.SerializedName
 
 data class MatchesTotalDTO(
-    @SerializedName("home") val home: Int,
-    @SerializedName("away") val away: Int,
-    @SerializedName("total") val total: Int
+    @SerializedName("home") val home: Float,
+    @SerializedName("away") val away: Float,
+    @SerializedName("total") val total: Float
 )
